condiguracion para deploy

+++ src/main/java/com/example/JAM23/demo/services/InscriptionService.java
@@ -1,7 +1,6 @@
 package com.example.JAM23.demo.services;
 
 import com.example.JAM23.demo.auth.User.User;
-import com.example.JAM23.demo.auth.UserRepository;
 import com.example.JAM23.demo.exception.customsExceptions.NotFoundException;
 import com.example.JAM23.demo.mappers.CourseMapper;
 import com.example.JAM23.demo.mappers.InscriptionMapper;
@@ -12,7 +11,6 @@ import com.example.JAM23.demo.model.dtos.inscriptions.InscriptionReadDto;
 import com.example.JAM23.demo.model.dtos.users.UserReadDto;
 import com.example.JAM23.demo.model.entities.CourseEntity;
 import com.example.JAM23.demo.model.entities.InscriptionEntity;
-import com.example.JAM23.demo.repositories.CourseRepository;
 import com.example.JAM23.demo.repositories.InscriptionRepository;
 import com.example.JAM23.demo.utils.Validator;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,25 +22,19 @@ import java.util.Optional;
 
 @Service
 public class InscriptionService {
-
-    @Autowired
-    private CourseService courseService;
-    @Autowired
-    private CourseRepository courseRepository;
-    @Autowired
-    private CourseMapper courseMapper;
-
     @Autowired
     private InscriptionRepository inscriptionRepository;
     @Autowired
     private InscriptionMapper inscriptionMapper;
 
     @Autowired
-    private UserRepository userRepository;
+    private CourseService courseService;
     @Autowired
-    private UserMapper userMapper;
+    private CourseMapper courseMapper;
     @Autowired
     private UserService userService;
+    @Autowired
+    private UserMapper userMapper;
 
     @Autowired
     private Validator validator;
@@ -67,17 +59,12 @@ public class InscriptionService {
                 .orElse(new InscriptionReadDto());
     }
     public List<UserReadDto> findAllInscriptedUsersByIdCourse(Integer idCourse){
-        List<User> inscriptedUsers = userRepository.findAllInscriptedUsersIdByIdCourse(idCourse);
-        List<UserReadDto> inscriptedUsersDtos = new ArrayList<>();
-        for (int i = 0; i<inscriptedUsers.size(); i++){
-            inscriptedUsersDtos.add(userMapper.userEntityTOUserReadDto(inscriptedUsers.get(i)));
-        }
-        return inscriptedUsersDtos;
+        return userService.findAllInscriptedUsersIdByIdCourse(idCourse);
     }
     public InscriptionReadDto findInscriptionById(Integer idInscription){
         InscriptionEntity insc = inscriptionRepository.findById(idInscription).orElseThrow(()-> new NotFoundException("Inscripcion no encontrada"));
-        UserReadDto user = userService.showUserById(insc.getUser().getId()); // getId_user_fk()
-        CourseReadDto course = courseService.showCourseById(insc.getCourse().getId()); //getId_course_fk()
+        User user = userService.getUserById(insc.getUser().getId()); // getId_user_fk()
+        CourseEntity course = courseService.getCourseById(insc.getCourse().getId()); //getId_course_fk()
 
         InscriptionReadDto inscription = new InscriptionReadDto();
         inscription.setId(insc.getId_insc());
@@ -96,13 +83,7 @@ public class InscriptionService {
         return  inscription;
     }
     public List<CourseReadDto> findAllCoursesInscriptedByUsername(String username){
-        List<CourseEntity> coursesEntities = courseRepository.findAllCoursesInscripted(username);
-        List<CourseReadDto> courseReadDtos = new ArrayList<>();
-
-        for(int i =0; i < coursesEntities.size() ; i++){
-            courseReadDtos.add(courseMapper.courseEntityTOCourseReadDto(coursesEntities.get(i)));
-        }
-        return courseReadDtos;
+        return courseService.findAllCoursesInscriptedByUsername(username);
     }
     public InscriptionReadDto editInscriptionById(Integer idInscription , InscriptionAddDto inscriptionAddDto){
         Integer id_user = inscriptionAddDto.getId_user();
@@ -113,12 +94,12 @@ public class InscriptionService {
         if(inscriptionEntity.isPresent()){
             if(id_user !=null) {
                 validator.existUserById(id_user);
-                User user = userRepository.findById(id_user).get();
+                User user = userService.getUserById(id_user);
                 inscriptionEntity.get().setUser(user);
             }
             if(id_course !=null) {
                 validator.existCourseById(id_course);
-                CourseEntity course = courseRepository.findById(id_course).get();
+                CourseEntity course = courseService.getCourseById(id_course);
                 inscriptionEntity.get().setCourse(course);
             }
             inscriptionRepository.save(inscriptionEntity.get());
@@ -127,7 +108,21 @@ public class InscriptionService {
         }
         return inscriptionMapper.inscriptionEntityTOInscriptionReadDto(inscriptionEntity.get());
     }
-    public void deleteInscriptionById(Integer idInscription){
-        inscriptionRepository.deleteById(idInscription);
+    public InscriptionReadDto deleteInscriptionById(Integer idInscription){
+        Optional<InscriptionEntity> insc = inscriptionRepository.findById(idInscription);
+        if (insc.isPresent()){
+            inscriptionRepository.deleteById(idInscription);
+            return inscriptionMapper.inscriptionEntityTOInscriptionReadDto(insc.get());
+        } else {
+            throw new NotFoundException("Inscripcion no encontrada con ID: "+idInscription);
+        }
+    }
+
+    public Integer getIdInscriptionByIdsCourseAndUser (Integer idCourse, Integer idUser){
+        try{
+            return inscriptionRepository.findByIdsCourseAndUser(idCourse , idUser).getId_insc();
+        } catch (Exception e){
+            throw new NotFoundException("Inscripcion al curso: "+idCourse+" del usuario: "+idUser+", no existe!");
+        }
     }
 }
